pyproject.toml
[tool.poetry]
name = "api"
version = "1.0.0"
description = ""
authors = ["Vladimir Melnichenko <melnichevv@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.3"
djangorestframework = "^3.13.1"
Markdown = "^3.3.6"
django-filter = "^21.1"
django-extensions = "^3.1.5"
djangorestframework-camel-case = "^1.3.0"
drf-yasg = "^1.20.0"
dj-database-url = "^0.5.0"
python-env-utils = "^0.4.1"
psycopg2-binary = "^2.9.3"
pre-commit = "^2.17.0"
django-cors-headers = "^3.11.0"
drf-jwt = "^1.19.2"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
ipython = "^8.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target_version = ['py39']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''